% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ivprobit.R
\docType{data}
\name{eco}
\alias{eco}
\title{Foreign-Exchange Derivatives Use By Large U.S. Bank Holding Com-panies (1996-2000).}
\format{A data frame with 794 rows and 22 variables}
\usage{
data(eco)
}
\description{
Foreign-Exchange Derivatives Use By Large U.S. Bank Holding Com-panies (1996-2000).
}
