% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ivprobit.R
\name{ivprobit}
\alias{ivprobit}
\title{Instrumental Variables Probit function}
\usage{
ivprobit(formula, data)
}
\arguments{
\item{formula}{y~x|y1|x2 whre y is the dichotomous l.h.s.,
               x is the r.h.s. exogenous variables, 
               y1 is the r.h.s. endogenous variables and 
               x2 is the complete set of instruments}

\item{data}{the dataframe}
}
\description{
Instrumental Variables Probit function
}
\examples{
############################################
# Fit the ivprobit model
############################################
# Load data
data(eco)
############################################
pro<-ivprobit(d2~ltass+roe+div|eqrat+bonus|ltass+roe+div+gap+cfa,eco)
summary(pro)

}

