% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ivprobit.R
\name{summary.ivprobit}
\alias{summary.ivprobit}
\title{Summary}
\usage{
\method{summary}{ivprobit}(object, ...)
}
\arguments{
\item{object}{is the object of the function}

\item{...}{not used}
}
\description{
Summary
}

