\name{ivProbit-package}
\alias{ivProbit-package}
\alias{ivProbit}
\docType{package}
\title{
Instrumental Variables Probit Model
}
\description{
In this package we compute the instrumental variables probit model using the generalized least squares estimator. This package comprises function that deal with endogeneity misspecification especially when correlation between regressors and error term is determinited and which produce inconsistent results.  
}
\details{
\tabular{ll}{
Package: \tab ivprobit\cr
Type: \tab Package\cr
Version: \tab 1.1\cr
Date: \tab 2018-02-21\cr
License: \tab GPL-3\cr
}
}

\author{
Taha Zaghdoudi

Taha Zaghdoudi <zedtaha@gmail.com>

}
