\name{ivProbit-package}
\alias{ivProbit-package}
\alias{ivProbit}
\docType{package}
\title{
Instrumental Variables Probit Model 
}
\description{
ivprobit fit an Instrumental variables probit model using the generalized least squares estimator 
}
\details{
\tabular{ll}{
Package: \tab ivprobit\cr
Type: \tab Package\cr
Version: \tab 1.1\cr
Date: \tab 2018-02-20\cr
License: \tab GPL-3\cr
}
}

\author{
Taha Zaghdoudi

Taha Zaghdoudi <zedtaha@gmail.com>

}
